backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx>=0.25.2
python-dotenv==1.0.1
ollama>=0.1.6
pytest>=8.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
prometheus-client==0.19.0
